refactor and edits in login/signup pages
